% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climb.R
\name{climb}
\alias{climb}
\title{CLIMB deconvolution}
\usage{
climb(
  sc,
  bulk,
  predict_expression = TRUE,
  ratio_cancer_cells = NA,
  up.lim = Inf,
  lambda = 0,
  norm_factor = 0.1
)
}
\arguments{
\item{sc}{ExpressionSet object containing the scRNA-seq reference dataset}

\item{bulk}{ExpressionSet object containing the mixtures to be deconvoluted}

\item{ratio_cancer_cells}{numeric vector, same size as number of mixtures. if we have data about the fraction of cancer cells present in mixture, this can improve deconvolution accuracy. If ratio_cancer_cells is provided, then CLIMB-BA method will be used. It is important that cancer cell-types in the scRNA-seq dataset contains the pattern "-like" to be recognized as cancer cells.}

\item{up.lim}{numeric scalar, impose a l-infinity norm to the linear model (upper bound for coefficient)}

\item{lambda}{Regularization factor}

\item{norm_factor}{if ratio_cancer_cells is provided, indicate strength of smooth normalization}
}
\description{
By using a scRNA-seq reference dataset (sc) and a bulk target dataset (bulk), CLIMB will generate cell-type abundance
and cell-type gene expression for each mixture given in the bulk data.
}
